% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apsrStars}
\alias{apsrStars}
\title{Custom summary functions for apsrtable stars}
\usage{
apsrStars(x, digits = max(3, getOption("digits") - 2),
  signif.stars = getOption("show.signif.stars"),
  signif.legend = signif.stars, dig.tst = max(1, min(5, digits - 1)),
  cs.ind = 1:k, tst.ind = k + 1, zap.ind = integer(0), P.values = NULL,
  has.Pvalue = nc >= 3 && substr(colnames(x)[nc], 1, 3) == "Pr(" || grepl("z",
  colnames(x)[nc]) == TRUE, eps.Pvalue = .Machine$double.eps,
  na.print = "NA", stars = "default", lev = 0.05, ...)
}
\arguments{
\item{x}{a matrix of coefficients}

\item{digits}{the number of digits to display}

\item{signif.stars}{the option for how to map stars onto p-values}

\item{signif.legend}{the legend for p-value mapping}

\item{dig.tst}{testing the number of digits}

\item{cs.ind}{unknown}

\item{tst.ind}{results for number of digits}

\item{zap.ind}{unknown}

\item{P.values}{which column are the p-values in the matrix of coefficients?}

\item{has.Pvalue}{are p-values present, true or false}

\item{eps.Pvalue}{unsure}

\item{na.print}{what character should represent NA}

\item{stars}{should stars be shown?}

\item{lev}{what is the significance level for p-values, numeric}

\item{...}{other options passed to the function}
}
\value{
A \code{summary} representation of a model object,
probably derived  from the object's own \code{summary} method.
}
\description{
Provides a programmatic way to include stars in tables
}
\author{
Michael Malecki <malecki at gmail.com>
}

