% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.matrix.VarCorr}
\alias{as.matrix.VarCorr}
\title{Extract VarCorr as a matrix from a mer-like object}
\usage{
\method{as.matrix}{VarCorr}(varc, useScale, digits)
}
\arguments{
\item{varc}{a VarCorr matrix from a merMod}

\item{useScale}{Rescale the matrix?}

\item{digits}{the number of digits to round to}
}
\description{
borrowed from package arm
}
\keyword{internal}

